Fix ios example (#24)
The same fix as in Android sdk:
https://github.com/fishjam-dev/android-client-sdk/pull/22

To reproduce: run the example ios app. The renegotiation event will be
sent before authRequest message due to a race condition.
This is more of a temporary solution to fix the example app quickly. A
more proper solution would be some code refactor.